Fixed issues with the refactoring that was done on azure service bus message system
